+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Le format est basé sur [Tenez un Changelog](https://keepachangelog.com/fr/1.0.0
 
 ## [À venir dans la prochaine version]
 ### Ajouté
+- Être notifié lorsqu'un user rejoint un projet
 - Ajouter une barre de recherche pour les faits marquant
 
 ## [0.35.2] - 2021-11-09